Adding reduce phase

+++ src/org/apache/gora/example/mapred/temporal/ReducerTempAggr.java
@@ -2,15 +2,15 @@ package org.apache.gora.example.mapred.temporal;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.mapreduce.Reducer;
-import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ReducerTempAggr<KEY> extends
-    Reducer<KEY, LongWritable, KEY, LongWritable> {
+    Reducer<KEY, FloatWritable, KEY, FloatWritable> {
 
-  private LongWritable result = new LongWritable();
+  private FloatWritable result = new FloatWritable();
   private static Logger Logger = LoggerFactory.getLogger(ReducerTempAggr.class);
 
   @Override
@@ -19,11 +19,11 @@ public class ReducerTempAggr<KEY> extends
     Logger.info(". . . Reducing . . .");
   }
 
-  public void reduce(KEY key, Iterable<LongWritable> values, Context context)
+  public void reduce(KEY key, Iterable<FloatWritable> values, Context context)
       throws IOException, InterruptedException {
     long sum = 0;
 
-    for (LongWritable val : values) {
+    for (FloatWritable val : values) {
       sum += val.get();
     }
     result.set(sum);
